+++ CHANGELOG.md
@@ -5,7 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## docx-wasm@0.0.276-rc29, rc28 (31. Oct, 2022)
+## docx-wasm@0.0.276-rc30 (1. Nov, 2022)
+
+- Use `None` if heightRule not found instead of `auto`
+
+## docx-wasm@0.0.276-rc29 (31. Oct, 2022)
 
 - escape hyperlink url(export)